# Faster cache by avoiding `try` call.