@schneems 2018-10-05
Even with #34068 this method is called about 3 times for every cache invocation. On CodeTriage's page, that's 158 calls per page load. 

Here's the new benchmark with #34068 merged

```ruby
require 'benchmark/ips'

nil_case = nil

Benchmark.ips do |x|
  x.report("nil &. ") { nil_case&.try(:foo) }
  x.report("nil try") { nil_case.try(:foo) }
  x.compare!
end
# Warming up --------------------------------------
#              nil &.    328.788k i/100ms
#              nil try   234.021k i/100ms
# Calculating -------------------------------------
#              nil &.      11.885M (± 3.2%) i/s -     59.511M in   5.013433s
#              nil try      5.758M (± 4.4%) i/s -     28.785M in   5.009582s

# Comparison:
#              nil &. : 11884504.8 i/s
#              nil try:  5757783.9 i/s - 2.06x  slower
```

The current method is 2x slower than this patch or you can say that this new method is ~50% faster (where 100% faster means it's instantaneous).

I did run some integration perf tests but this is a really tiny change. I'm seeing improved time, but also high variance. I need to schedule some time to do a "proper" test that can maybe provide more insights.