@rafaelfranca 2018-10-05
Given the integration perf tests show this as a really tiny change this fall in the micro benchmark category that we continuously reject across the framework.

As a reader of the code this core doesn't make any sense to me. If I'm calling `try` that knows how to deal with `nil` why I would check for new again?

This is equivalent to `foo && foo.try(:something)`. This code is something we would without questioning change to remove the redundant `&&` operation. I don't think we should include in the framework.

Other point to consider this falls in the category of things the VM should be optimizing for me, so instead of changing the way I wrote my code to get a 1% faster application I'd spend time on optimizing the VM.