@schneems 2018-10-06
Re-ran my scripts and it looks like the real world impact to CodeTriage is about 0.14% faster page load.